+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - [495](https://github.com/vegaprotocol/data-node/pull/495) - Remove deprecated `PositionState` event handling, general fixes to `SettlePosition` event handling
 - [498](https://github.com/vegaprotocol/data-node/issues/498) - Transaction event broker
 - [521](https://github.com/vegaprotocol/data-node/issues/521) - Refactor margin levels to use account id 
+- [533](https://github.com/vegaprotocol/data-node/issues/533) - Order data retention 
 
 - [](https://github.com/vegaprotocol/data-node/pull/xxx) -